+++ testing/node-driver/src/main/kotlin/net/corda/testing/node/InMemoryMessagingNetwork.kt
@@ -1,45 +1,36 @@
 package net.corda.testing.node
 
+import net.corda.core.CordaInternal
 import net.corda.core.DoNotImplement
 import net.corda.core.crypto.CompositeKey
 import net.corda.core.identity.CordaX500Name
 import net.corda.core.identity.Party
 import net.corda.core.identity.PartyAndCertificate
-import net.corda.core.internal.ThreadBox
 import net.corda.core.internal.concurrent.openFuture
 import net.corda.core.messaging.AllPossibleRecipients
 import net.corda.core.messaging.MessageRecipientGroup
 import net.corda.core.messaging.MessageRecipients
 import net.corda.core.messaging.SingleMessageRecipient
-import net.corda.core.node.services.PartyInfo
 import net.corda.core.serialization.CordaSerializable
 import net.corda.core.serialization.SingletonSerializeAsToken
 import net.corda.core.utilities.ByteSequence
-import net.corda.core.utilities.OpaqueBytes
 import net.corda.core.utilities.getOrThrow
 import net.corda.core.utilities.trace
-import net.corda.node.services.messaging.*
-import net.corda.node.services.statemachine.DeduplicationId
-import net.corda.node.services.statemachine.ExternalEvent
-import net.corda.node.services.statemachine.SenderDeduplicationId
-import net.corda.node.utilities.AffinityExecutor
-import net.corda.testing.node.internal.InMemoryMessage
-import net.corda.testing.node.internal.InternalMockMessagingService
+import net.corda.node.services.messaging.Message
+import net.corda.testing.node.internal.MockNodeMessagingService
 import org.apache.activemq.artemis.utils.ReusableLatch
 import org.slf4j.LoggerFactory
 import rx.Observable
 import rx.subjects.PublishSubject
 import java.time.Duration
-import java.time.Instant
 import java.util.*
 import java.util.concurrent.LinkedBlockingQueue
 import javax.annotation.concurrent.ThreadSafe
 import kotlin.concurrent.schedule
-import kotlin.concurrent.thread
 
 /**
- * An in-memory network allows you to manufacture [InternalMockMessagingService]s for a set of participants. Each
- * [InternalMockMessagingService] maintains a queue of messages it has received, and a background thread that dispatches
+ * An in-memory network allows you to manufacture [MockNodeMessagingService]s for a set of participants. Each
+ * [MockNodeMessagingService] maintains a queue of messages it has received, and a background thread that dispatches
  * messages one by one to registered handlers. Alternatively, a messaging system may be manually pumped, in which
  * case no thread is created and a caller is expected to force delivery one at a time (this is useful for unit
  * testing).
@@ -57,16 +48,16 @@ class InMemoryMessagingNetwork private constructor(
         private const val MESSAGES_LOG_NAME = "messages"
         private val log = LoggerFactory.getLogger(MESSAGES_LOG_NAME)
 
-        internal fun create(
-                sendManuallyPumped: Boolean,
-                servicePeerAllocationStrategy: ServicePeerAllocationStrategy = InMemoryMessagingNetwork.ServicePeerAllocationStrategy.Random(),
-                messagesInFlight: ReusableLatch = ReusableLatch()): InMemoryMessagingNetwork {
+        @CordaInternal
+        internal fun create(sendManuallyPumped: Boolean,
+                            servicePeerAllocationStrategy: ServicePeerAllocationStrategy = InMemoryMessagingNetwork.ServicePeerAllocationStrategy.Random(),
+                            messagesInFlight: ReusableLatch = ReusableLatch()): InMemoryMessagingNetwork {
             return InMemoryMessagingNetwork(sendManuallyPumped, servicePeerAllocationStrategy, messagesInFlight)
         }
     }
 
     private var counter = 0   // -1 means stopped.
-    private val handleEndpointMap = HashMap<PeerHandle, InMemoryMessaging>()
+    private val handleEndpointMap = HashMap<PeerHandle, MockNodeMessagingService>()
 
     /** A class which represents a message being transferred from sender to recipients, within the [InMemoryMessageNetwork]. **/
     @CordaSerializable
@@ -107,40 +98,35 @@ class InMemoryMessagingNetwork private constructor(
     /** A stream of (sender, message, recipients) triples containing messages once they have been received. */
     val receivedMessages: Observable<MessageTransfer>
         get() = _receivedMessages
-    internal val endpoints: List<InternalMockMessagingService> @Synchronized get() = handleEndpointMap.values.toList()
+    internal val endpoints: List<MockNodeMessagingService>
+        @CordaInternal
+        @Synchronized
+        get() = handleEndpointMap.values.toList()
     /** Get a [List] of all the [MockMessagingService] endpoints **/
-    val endpointsExternal: List<MockMessagingService> @Synchronized get() = handleEndpointMap.values.map { MockMessagingService.createMockMessagingService(it) }.toList()
+    val endpointsExternal: List<MockMessagingService>
+        @Synchronized
+        get() = handleEndpointMap.values.map { MockMessagingService.createMockMessagingService(it) }.toList()
 
-    /**
-     * Creates a node at the given address: useful if you want to recreate a node to simulate a restart.
-     *
-     * @param manuallyPumped if set to true, then you are expected to call [InMemoryMessaging.pumpReceive]
-     * in order to cause the delivery of a single message, which will occur on the thread of the caller. If set to false
-     * then this class will set up a background thread to deliver messages asynchronously, if the handler specifies no
-     * executor.
-     * @param id the numeric ID to use, e.g. set to whatever ID the node used last time.
-     * @param description text string that identifies this node for message logging (if is enabled) or null to autogenerate.
-     */
-    internal fun createNodeWithID(
-            manuallyPumped: Boolean,
-            id: Int,
-            executor: AffinityExecutor,
-            description: CordaX500Name = CordaX500Name(organisation = "In memory node $id", locality = "London", country = "UK")
-    ): InternalMockMessagingService {
-        val peerHandle = PeerHandle(id, description)
-        peersMapping[peerHandle.name] = peerHandle // Assume that the same name - the same entity in MockNetwork.
+    @CordaInternal
+    internal fun getPeer(name: CordaX500Name): PeerHandle? = peersMapping[name]
+
+    @CordaInternal
+    internal fun initPeer(messagingService: MockNodeMessagingService): MockNodeMessagingService? {
+        peersMapping[messagingService.myAddress.name] = messagingService.myAddress // Assume that the same name - the same entity in MockNetwork.
         return synchronized(this) {
-            val node = InMemoryMessaging(manuallyPumped, peerHandle, executor)
-            val oldNode = handleEndpointMap.put(peerHandle, node)
-            if (oldNode != null) {
-                node.inheritPendingRedelivery(oldNode)
-            }
-            node
+            handleEndpointMap.put(messagingService.myAddress, messagingService)
         }
     }
 
-    internal fun onNotaryIdentity(node: InternalMockMessagingService, notaryService: PartyAndCertificate?) {
-        val peerHandle = (node as InMemoryMessaging).peerHandle
+    @CordaInternal
+    internal fun onMessageTransfer(transfer: MessageTransfer) {
+        _receivedMessages.onNext(transfer)
+        messagesInFlight.countDown()
+    }
+
+    @CordaInternal
+    internal fun addNotaryIdentity(node: MockNodeMessagingService, notaryService: PartyAndCertificate?) {
+        val peerHandle = node.myAddress
         notaryService?.let { if (it.owningKey !is CompositeKey) peersMapping[it.name] = peerHandle }
         val serviceHandles = notaryService?.let { listOf(DistributedServiceHandle(it.party)) }
                 ?: emptyList() //TODO only notary can be distributed?
@@ -161,22 +147,31 @@ class InMemoryMessagingNetwork private constructor(
     private var latencyCalculator: LatencyCalculator? = null
     private val timer = Timer()
 
-    @Synchronized
-    private fun msgSend(from: InMemoryMessaging, message: Message, recipients: MessageRecipients) {
-        messagesInFlight.countUp()
-        messageSendQueue += MessageTransfer.createMessageTransfer(from.myAddress, message, recipients)
+    @CordaInternal
+    internal fun msgSend(from: MockNodeMessagingService, message: Message, recipients: MessageRecipients) {
+        synchronized(this) {
+            messagesInFlight.countUp()
+            messageSendQueue += MessageTransfer.createMessageTransfer(from.myAddress, message, recipients)
+        }
+        if (!sendManuallyPumped) {
+            pumpSend(false)
+        }
     }
 
+    @CordaInternal
     @Synchronized
-    private fun netNodeHasShutdown(peerHandle: PeerHandle) {
+    internal fun netNodeHasShutdown(peerHandle: PeerHandle) {
         val endpoint = handleEndpointMap[peerHandle]
-        if (!(endpoint?.hasPendingDeliveries() ?: false)) {
+        if (endpoint?.hasPendingDeliveries() != true) {
             handleEndpointMap.remove(peerHandle)
         }
     }
 
+    @CordaInternal
     @Synchronized
-    private fun getQueueForPeerHandle(recipients: PeerHandle) = messageReceiveQueues.getOrPut(recipients) { LinkedBlockingQueue() }
+    internal fun getQueueForPeerHandle(recipients: PeerHandle): LinkedBlockingQueue<MessageTransfer> {
+        return messageReceiveQueues.getOrPut(recipients) { LinkedBlockingQueue() }
+    }
 
     @Synchronized
     private fun getQueuesForServiceHandle(recipients: DistributedServiceHandle): List<LinkedBlockingQueue<MessageTransfer>> {
@@ -276,30 +271,6 @@ class InMemoryMessagingNetwork private constructor(
         return transfer
     }
 
-    /**
-     * When a new message handler is added, this implies we have started a new node.  The add handler logic uses this to
-     * push back any un-acknowledged messages for this peer onto the head of the queue (rather than the tail) to maintain message
-     * delivery order.  We push them back because their consumption was not complete and a restarted node would
-     * see them re-delivered if this was Artemis.
-     */
-    @Synchronized
-    private fun unPopMessages(transfers: Collection<MessageTransfer>, us: PeerHandle) {
-        messageReceiveQueues.compute(us) { _, existing ->
-            if (existing == null) {
-                LinkedBlockingQueue<MessageTransfer>().apply {
-                    addAll(transfers)
-                }
-            } else {
-                existing.apply {
-                    val drained = mutableListOf<MessageTransfer>()
-                    existing.drainTo(drained)
-                    existing.addAll(transfers)
-                    existing.addAll(drained)
-                }
-            }
-        }
-    }
-
     private fun pumpSendInternal(transfer: MessageTransfer) {
         when (transfer.recipients) {
             is PeerHandle -> getQueueForPeerHandle(transfer.recipients).add(transfer)
@@ -311,35 +282,25 @@ class InMemoryMessagingNetwork private constructor(
             is AllPossibleRecipients -> {
                 // This means all possible recipients _that the network knows about at the time_, not literally everyone
                 // who joins into the indefinite future.
-                for (handle in handleEndpointMap.keys)
-                    getQueueForPeerHandle(handle).add(transfer)
+                synchronized(this) {
+                    for (handle in handleEndpointMap.keys) {
+                        getQueueForPeerHandle(handle).add(transfer)
+                    }
+                }
             }
             else -> throw IllegalArgumentException("Unknown type of recipient handle")
         }
         _sentMessages.onNext(transfer)
     }
 
-    private data class InMemoryReceivedMessage(override val topic: String,
-                                               override val data: ByteSequence,
-                                               override val platformVersion: Int,
-                                               override val uniqueMessageId: DeduplicationId,
-                                               override val debugTimestamp: Instant,
-                                               override val peer: CordaX500Name,
-                                               override val senderUUID: String? = null,
-                                               override val senderSeqNo: Long? = null,
-                                               /** Note this flag is never set in the in memory network. */
-                                               override val isSessionInit: Boolean = false) : ReceivedMessage {
-
-        override val additionalHeaders: Map<String, String> = emptyMap()
-    }
-
     /**
      * A class that provides an abstraction over the nodes' messaging service that also contains the ability to
      * receive messages from the queue for testing purposes.
      */
-    class MockMessagingService private constructor(private val messagingService: InternalMockMessagingService) {
+    class MockMessagingService private constructor(private val messagingService: MockNodeMessagingService) {
         companion object {
-            internal fun createMockMessagingService(messagingService: InternalMockMessagingService): MockMessagingService {
+            @CordaInternal
+            internal fun createMockMessagingService(messagingService: MockNodeMessagingService): MockMessagingService {
                 return MockMessagingService(messagingService)
             }
         }
@@ -352,197 +313,4 @@ class InMemoryMessagingNetwork private constructor(
          */
         fun pumpReceive(block: Boolean): InMemoryMessagingNetwork.MessageTransfer? = messagingService.pumpReceive(block)
     }
-
-    @ThreadSafe
-    private inner class InMemoryMessaging(private val manuallyPumped: Boolean,
-                                          val peerHandle: PeerHandle,
-                                          private val executor: AffinityExecutor) : SingletonSerializeAsToken(), InternalMockMessagingService {
-        private inner class Handler(val topicSession: String, val callback: MessageHandler) : MessageHandlerRegistration
-
-        @Volatile
-        private var running = true
-
-        private inner class InnerState {
-            val handlers: MutableList<Handler> = ArrayList()
-            val pendingRedelivery = LinkedHashSet<MessageTransfer>()
-        }
-
-        private val state = ThreadBox(InnerState())
-        private val processedMessages: MutableSet<DeduplicationId> = Collections.synchronizedSet(HashSet<DeduplicationId>())
-
-        override val myAddress: PeerHandle get() = peerHandle
-        override val ourSenderUUID: String = UUID.randomUUID().toString()
-
-        private val backgroundThread = if (manuallyPumped) null else
-            thread(isDaemon = true, name = "In-memory message dispatcher") {
-                while (!Thread.currentThread().isInterrupted) {
-                    try {
-                        pumpReceiveInternal(true)
-                    } catch (e: InterruptedException) {
-                        break
-                    }
-                }
-            }
-
-        override fun getAddressOfParty(partyInfo: PartyInfo): MessageRecipients {
-            return when (partyInfo) {
-                is PartyInfo.SingleNode -> peersMapping[partyInfo.party.name]
-                        ?: throw IllegalArgumentException("No StartedMockNode for party ${partyInfo.party.name}")
-                is PartyInfo.DistributedNode -> DistributedServiceHandle(partyInfo.party)
-            }
-        }
-
-        override fun addMessageHandler(topic: String, callback: MessageHandler): MessageHandlerRegistration {
-            check(running)
-            val (handler, transfers) = state.locked {
-                val handler = Handler(topic, callback).apply { handlers.add(this) }
-                val pending = ArrayList<MessageTransfer>()
-                pending.addAll(pendingRedelivery)
-                pendingRedelivery.clear()
-                Pair(handler, pending)
-            }
-
-            unPopMessages(transfers, peerHandle)
-            return handler
-        }
-
-        fun inheritPendingRedelivery(other: InMemoryMessaging) {
-            state.locked {
-                pendingRedelivery.addAll(other.state.locked { pendingRedelivery })
-            }
-        }
-
-        override fun removeMessageHandler(registration: MessageHandlerRegistration) {
-            check(running)
-            state.locked { check(handlers.remove(registration as Handler)) }
-        }
-
-        override fun send(message: Message, target: MessageRecipients, sequenceKey: Any) {
-            check(running)
-            msgSend(this, message, target)
-            if (!sendManuallyPumped) {
-                pumpSend(false)
-            }
-        }
-
-        override fun send(addressedMessages: List<MessagingService.AddressedMessage>) {
-            for ((message, target, sequenceKey) in addressedMessages) {
-                send(message, target, sequenceKey)
-            }
-        }
-
-        override fun close() {
-            if (backgroundThread != null) {
-                backgroundThread.interrupt()
-                backgroundThread.join()
-            }
-            running = false
-            netNodeHasShutdown(peerHandle)
-        }
-
-        /** Returns the given (topic & session, data) pair as a newly created message object. */
-        override fun createMessage(topic: String, data: ByteArray, deduplicationId: SenderDeduplicationId, additionalHeaders: Map<String, String>): Message {
-            return InMemoryMessage(topic, OpaqueBytes(data), deduplicationId.deduplicationId, senderUUID = deduplicationId.senderUUID)
-        }
-
-        /**
-         * Delivers a single message from the internal queue. If there are no messages waiting to be delivered and block
-         * is true, waits until one has been provided on a different thread via send. If block is false, the return
-         * result indicates whether a message was delivered or not.
-         *
-         * @return the message that was processed, if any in this round.
-         */
-        override fun pumpReceive(block: Boolean): MessageTransfer? {
-            check(manuallyPumped)
-            check(running)
-            executor.flush()
-            try {
-                return pumpReceiveInternal(block)
-            } finally {
-                executor.flush()
-            }
-        }
-
-        /**
-         * Get the next transfer, and matching queue, that is ready to handle. Any pending transfers without handlers
-         * are placed into `pendingRedelivery` to try again later.
-         *
-         * @param block if this should block until a message it can process.
-         */
-        private fun getNextQueue(q: LinkedBlockingQueue<MessageTransfer>, block: Boolean): Pair<MessageTransfer, List<Handler>>? {
-            var deliverTo: List<Handler>? = null
-            // Pop transfers off the queue until we run out (and are not blocking), or find something we can process
-            while (deliverTo == null) {
-                val transfer = (if (block) q.take() else q.poll()) ?: return null
-                deliverTo = state.locked {
-                    val matchingHandlers = handlers.filter { it.topicSession.isBlank() || transfer.message.topic == it.topicSession }
-                    if (matchingHandlers.isEmpty()) {
-                        // Got no handlers for this message yet. Keep the message around and attempt redelivery after a new
-                        // handler has been registered. The purpose of this path is to make unit tests that have multi-threading
-                        // reliable, as a sender may attempt to send a message to a receiver that hasn't finished setting
-                        // up a handler for yet. Most unit tests don't run threaded, but we want to test true parallelism at
-                        // least sometimes.
-                        log.warn("Message to ${transfer.message.topic} could not be delivered")
-                        pendingRedelivery.add(transfer)
-                        null
-                    } else {
-                        matchingHandlers
-                    }
-                }
-                if (deliverTo != null) {
-                    return Pair(transfer, deliverTo)
-                }
-            }
-            return null
-        }
-
-        private fun pumpReceiveInternal(block: Boolean): MessageTransfer? {
-            val q = getQueueForPeerHandle(peerHandle)
-            val (transfer, deliverTo) = getNextQueue(q, block) ?: return null
-            if (transfer.message.uniqueMessageId !in processedMessages) {
-                executor.execute {
-                    for (handler in deliverTo) {
-                        try {
-                            val receivedMessage = transfer.toReceivedMessage()
-                            state.locked { pendingRedelivery.add(transfer) }
-                            handler.callback(receivedMessage, handler, InMemoryDeduplicationHandler(receivedMessage, transfer))
-                        } catch (e: Exception) {
-                            log.error("Caught exception in handler for $this/${handler.topicSession}", e)
-                        }
-                    }
-                    _receivedMessages.onNext(transfer)
-                    messagesInFlight.countDown()
-                }
-            } else {
-                log.info("Drop duplicate message ${transfer.message.uniqueMessageId}")
-            }
-            return transfer
-        }
-
-        private fun MessageTransfer.toReceivedMessage(): ReceivedMessage = InMemoryReceivedMessage(
-                message.topic,
-                OpaqueBytes(message.data.bytes.copyOf()), // Kryo messes with the buffer so give each client a unique copy
-                1,
-                message.uniqueMessageId,
-                message.debugTimestamp,
-                sender.name)
-
-        private inner class InMemoryDeduplicationHandler(override val receivedMessage: ReceivedMessage, val transfer: MessageTransfer) : DeduplicationHandler, ExternalEvent.ExternalMessageEvent {
-            override val externalCause: ExternalEvent
-                get() = this
-            override val deduplicationHandler: DeduplicationHandler
-                get() = this
-
-            override fun afterDatabaseTransaction() {
-                this@InMemoryMessaging.state.locked { pendingRedelivery.remove(transfer) }
-            }
-
-            override fun insideDatabaseTransaction() {
-                processedMessages += transfer.message.uniqueMessageId
-            }
-        }
-
-        fun hasPendingDeliveries(): Boolean = state.locked { pendingRedelivery.isNotEmpty() }
-    }
 }
-
